requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.2.1
biopython==1.83
# pdfminer.six is optional; uncomment if you need arXiv PDF extraction
# pdfminer.six==20231228
